Improve readability for website/docs/configuration
This commit changes "Creating a shareable config will extract this config to an Alfred config. In this case, it makes sense to call the config `alfred-config-web-app`" to "Creating a shareable config will extract this config into an Alfred config. In this case, it makes sense to call the config `alfred-config-web-app`".

Co-authored-by: Amila Welihinda <amilajack@gmail.com>