Close #335 by fixing typos and removing instances of passive voice (#336)
* close #335 by fixing typos to improve readability

This commit also changes a couple instances of the passive voice to "Never use the passive where you can use the active", as stated in this repository's CONTRIBUTING.md file.
https://github.com/amilajack/alfred/blob/master/CONTRIBUTING.md#guidelines-for-writing-docs

* change "hello work" to "hello world" in 2 READMEs

This commit changes "hello work" to "hello world" in examples/README.md and README.md.

* Improve readability of website/docs/configuration

This commit changes "The directory, relative to the project directory, to write config files" to "The path (relative to project directory) that config files are written to" to improve readability.

Co-authored-by: Amila Welihinda <amilajack@gmail.com>

* Improve readability for website/docs/configuration

This commit changes "Creating a shareable config will extract this config to an Alfred config. In this case, it makes sense to call the config `alfred-config-web-app`" to "Creating a shareable config will extract this config into an Alfred config. In this case, it makes sense to call the config `alfred-config-web-app`".

Co-authored-by: Amila Welihinda <amilajack@gmail.com>

Co-authored-by: Amila Welihinda <amilajack@gmail.com>